improved option chain data

+++ Trading.Domain/Models/TradeSignal.cs
@@ -1,9 +1,9 @@
 ﻿public class TradeSignal
 {
     public string Signal { get; set; }
-    public double Strike { get; set; }
+    public decimal Strike { get; set; }
     public string Symbol { get; set; }
-    public double Confidence { get; set; }
+    public decimal Confidence { get; set; }
     public decimal Entry { get; set; }
     public decimal StopLoss { get; set; }
     public decimal Target { get; set; }
